+++ Sandbox/GLSLBook/GLSLBookPlugins.cpp
@@ -6,6 +6,8 @@
 
 #include "Core/RHI/RenderSystem.h"
 #include "GLSLBook/PostEffectPipeline.h"
+#include "GLSLBook/ImageProcessing/EdgePipeline.h"
+#include "GLSLBook/ImageProcessing/BlurPipeline.h"
 
 using namespace Zelo::Core::RHI;
 
@@ -15,7 +17,7 @@ const std::string &EdgePipelinePlugin::getName() const {
 }
 
 void EdgePipelinePlugin::install() {
-    auto renderPipeline = std::make_unique<PostEffectPipeline>();
+    auto renderPipeline = std::make_unique<EdgePipeline>();
     renderPipeline->initialize();
     RenderSystem::getSingletonPtr()->setRenderPipeline(std::move(renderPipeline));
 }
@@ -29,5 +31,28 @@ void EdgePipelinePlugin::shutdown() {
 }
 
 void EdgePipelinePlugin::uninstall() {
+    RenderSystem::getSingletonPtr()->resetRenderPipeline();
+}
+
+const std::string &BlurPipelinePlugin::getName() const {
+    static std::string s = "PostEffect";
+    return s;
+}
+
+void BlurPipelinePlugin::install() {
+    auto renderPipeline = std::make_unique<BlurPipeline>();
+    renderPipeline->initialize();
+    RenderSystem::getSingletonPtr()->setRenderPipeline(std::move(renderPipeline));
+}
+
+void BlurPipelinePlugin::initialise() {
+    // do nothing
+}
+
+void BlurPipelinePlugin::shutdown() {
+    // do nothing
+}
+
+void BlurPipelinePlugin::uninstall() {
     RenderSystem::getSingletonPtr()->resetRenderPipeline();
 }
