+++ Script/Lua/main_initialize.lua
@@ -92,6 +92,7 @@ TheFrontEnd:LoadPanel(MaterialEditor, "Material Editor", SHOW_ALL_PANEL_AT_INIT)
 UI:ResetLayout()
 
 postEffectPlugin = EdgePipelinePlugin.new()
+postEffectPlugin = BlurPipelinePlugin.new()
 install(postEffectPlugin)
 
 -- scene
